chore: Refactor structure
* No more complex Python import hell
