--- src/scaict_uwu/config/factory.py
@@ -1,11 +1,0 @@
-"""
-Config factory
-"""
-
-# Local imports
-from .config import Config
-
-
-class ConfigFactory:
-    def get(self) -> Config:
-        return Config()
