--- src/scaict_uwu/config/names.py
@@ -1,7 +1,0 @@
-"""
-Config names
-"""
-
-
-class ConfigNames:
-    pass
