--- src/scaict_uwu/config/schema.py
@@ -1,7 +1,0 @@
-"""
-Config schema
-"""
-
-
-class ConfigSchema:
-    pass
