--- src/scaict_uwu/libs/services/errors/cannot_replace_active_service_error.py
@@ -1,9 +1,0 @@
-class CannotReplaceActiveServiceError(Exception):
-    """
-    Error thrown when trying to replace an already active service.
-    """
-
-    def __init__(self, *, name: str):
-        formatted_message = f"Cannot replace an active service: {name}"
-
-        super().__init__(formatted_message)
