--- src/scaict_uwu/libs/services/errors/no_such_service_error.py
@@ -1,9 +1,0 @@
-class NoSuchServiceError(Exception):
-    """
-    Error thrown when the requested service is not known.
-    """
-
-    def __init__(self, *, name: str):
-        formatted_message = f"No such service: {name}"
-
-        super().__init__(formatted_message)
