+++ src/scaict_uwu/maintenance/scripts.py
@@ -1,0 +1,28 @@
+"""
+Maintenance scripts
+"""
+
+# Standard imports
+# Prevents runtime evaluation of type hints
+from __future__ import annotations
+
+# Local imports
+from scaict_uwu.maintenance.maintenance import MaintenanceScript
+
+
+class MaintenanceScriptUpdate(MaintenanceScript):
+    """
+    Maintenance script to run database schema updates.
+
+    Run all updaters.
+
+    This is used when the database schema is modified and we need to apply patches.
+    """
+
+    def __init__(self) -> None:
+        super().__init__()
+
+        self.add_description("Database updater.")
+
+    def execute(self) -> bool:
+        return True
