--- src/scaict_uwu/user/factory.py
@@ -1,9 +1,0 @@
-"""
-User factory
-"""
-
-
-class UserFactory:
-    """
-    User factory
-    """
