--- src/scaict_uwu/user/user.py
@@ -1,9 +1,0 @@
-"""
-User
-"""
-
-
-class User:
-    """
-    User data class
-    """
