+++ src/scaict_uwu/website.py
@@ -2,6 +2,10 @@
 Entry point for the SCAICT store website using Flask run.
 """
 
+# Standard imports
+# Prevents runtime evaluation of type hints
+from __future__ import annotations
+
 
 def run() -> None:
     """
